requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
click==8.0.1
    # via -r requirements.in
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
pygments==2.10.0
    # via rich
python-nmap==0.6.4
    # via -r requirements.in
rich==10.7.0
    # via -r requirements.in
